refactoring async bulk controllers
