+++ app/code/Magento/WebapiAsync/Controller/Rest/AsynchronousRequestProcessor.php
@@ -11,46 +11,51 @@ namespace Magento\WebapiAsync\Controller\Rest;
 use Magento\Framework\Exception\BulkException;
 use Magento\Webapi\Controller\Rest\RequestProcessorInterface;
 use Magento\Framework\Webapi\Rest\Response as RestResponse;
-use Magento\Webapi\Controller\Rest\InputParamsResolver;
+use Magento\WebapiAsync\Controller\Rest\Asynchronous\InputParamsResolver;
 use Magento\AsynchronousOperations\Model\MassSchedule;
 use Magento\AsynchronousOperations\Model\ConfigInterface as WebApiAsyncConfig;
 use Magento\Framework\Reflection\DataObjectProcessor;
 use Magento\AsynchronousOperations\Api\Data\AsyncResponseInterfaceFactory;
 use Magento\AsynchronousOperations\Api\Data\AsyncResponseInterface;
 
+/**
+ * Responsible for dispatching single and bulk requests.
+ * Single requests dispatching represented by this class.
+ * Bulk requests dispatching represented by virtualType of this class.
+ */
 class AsynchronousRequestProcessor implements RequestProcessorInterface
 {
     const PROCESSOR_PATH = "/^\\/async(\\/V.+)/";
+    const BULK_PROCESSOR_PATH = "/^\\/async\/bulk(\\/V.+)/";
 
     /**
      * @var \Magento\Framework\Webapi\Rest\Response
      */
     private $response;
-
     /**
-     * @var InputParamsResolver
+     * @var \Magento\WebapiAsync\Controller\Rest\Asynchronous\InputParamsResolver
      */
     private $inputParamsResolver;
-
     /**
      * @var MassSchedule
      */
     private $asyncBulkPublisher;
-
     /**
      * @var WebApiAsyncConfig
      */
     private $webapiAsyncConfig;
-
     /**
      * @var \Magento\Framework\Reflection\DataObjectProcessor
      */
     private $dataObjectProcessor;
-
     /**
      * @var AsyncResponseInterfaceFactory
      */
     private $asyncResponseFactory;
+    /**
+     * @var string Regex pattern
+     */
+    private $processorPath;
 
     /**
      * Initialize dependencies.
@@ -61,6 +66,7 @@ class AsynchronousRequestProcessor implements RequestProcessorInterface
      * @param WebapiAsyncConfig $webapiAsyncConfig
      * @param DataObjectProcessor $dataObjectProcessor
      * @param AsyncResponseInterfaceFactory $asyncResponse
+     * @param string $processorPath
      */
     public function __construct(
         RestResponse $response,
@@ -68,7 +74,8 @@ class AsynchronousRequestProcessor implements RequestProcessorInterface
         MassSchedule $asyncBulkPublisher,
         WebApiAsyncConfig $webapiAsyncConfig,
         DataObjectProcessor $dataObjectProcessor,
-        AsyncResponseInterfaceFactory $asyncResponse
+        AsyncResponseInterfaceFactory $asyncResponse,
+        $processorPath = self::PROCESSOR_PATH
     ) {
         $this->response = $response;
         $this->inputParamsResolver = $inputParamsResolver;
@@ -76,6 +83,7 @@ class AsynchronousRequestProcessor implements RequestProcessorInterface
         $this->webapiAsyncConfig = $webapiAsyncConfig;
         $this->dataObjectProcessor = $dataObjectProcessor;
         $this->asyncResponseFactory = $asyncResponse;
+        $this->processorPath = $processorPath;
     }
 
     /**
@@ -84,12 +92,12 @@ class AsynchronousRequestProcessor implements RequestProcessorInterface
     public function process(\Magento\Framework\Webapi\Rest\Request $request)
     {
         $path = $request->getPathInfo();
-        $path = preg_replace(self::PROCESSOR_PATH, "$1", $path);
+        $path = preg_replace($this->processorPath, "$1", $path);
         $request->setPathInfo(
             $path
         );
 
-        $entitiesParamsArray = [$this->inputParamsResolver->resolve()];
+        $entitiesParamsArray = $this->inputParamsResolver->resolve();
         $topicName = $this->getTopicName($request);
 
         try {
@@ -133,7 +141,19 @@ class AsynchronousRequestProcessor implements RequestProcessorInterface
             return false;
         }
 
-        if (preg_match(self::PROCESSOR_PATH, $request->getPathInfo()) === 1) {
+        if (preg_match($this->processorPath, $request->getPathInfo()) === 1) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * @param \Magento\Framework\Webapi\Rest\Request $request
+     * @return bool
+     */
+    public function isBulk(\Magento\Framework\Webapi\Rest\Request $request)
+    {
+        if (preg_match(self::BULK_PROCESSOR_PATH, $request->getPathInfo()) === 1) {
             return true;
         }
         return false;
