+++ app/code/Magento/WebapiAsync/Controller/Rest/AsynchronousSchemaRequestProcessor.php
@@ -15,34 +15,40 @@ use Magento\Webapi\Controller\Rest\RequestProcessorInterface;
 
 class AsynchronousSchemaRequestProcessor implements RequestProcessorInterface
 {
-
     /**
      * Path for accessing Async Rest API schema
      */
     const PROCESSOR_PATH = 'async/schema';
+    const BULK_PROCESSOR_PATH = 'async/bulk/schema';
 
     /**
      * @var \Magento\Webapi\Model\Rest\Swagger\Generator
      */
     private $swaggerGenerator;
-
     /**
      * @var \Magento\Framework\Webapi\Rest\Response
      */
     private $response;
+    /**
+     * @var string
+     */
+    private $processorPath;
 
     /**
      * Initial dependencies
      *
      * @param \Magento\Webapi\Model\Rest\Swagger\Generator $swaggerGenerator
      * @param \Magento\Framework\Webapi\Rest\Response $response
+     * @param string $processorPath
      */
     public function __construct(
         Generator $swaggerGenerator,
-        RestResponse $response
+        RestResponse $response,
+        $processorPath = self::PROCESSOR_PATH
     ) {
         $this->swaggerGenerator = $swaggerGenerator;
         $this->response = $response;
+        $this->processorPath = $processorPath;
     }
 
     /**
@@ -68,7 +74,19 @@ class AsynchronousSchemaRequestProcessor implements RequestProcessorInterface
      */
     public function canProcess(\Magento\Framework\Webapi\Rest\Request $request)
     {
-        if (strpos(ltrim($request->getPathInfo(), '/'), self::PROCESSOR_PATH) === 0) {
+        if (strpos(ltrim($request->getPathInfo(), '/'), $this->processorPath) === 0) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * @param \Magento\Framework\Webapi\Rest\Request $request
+     * @return bool
+     */
+    public function isBulk(\Magento\Framework\Webapi\Rest\Request $request)
+    {
+        if (strpos(ltrim($request->getPathInfo(), '/'), self::BULK_PROCESSOR_PATH) === 0) {
             return true;
         }
         return false;
