+++ app/code/Magento/WebapiAsync/etc/di.xml
@@ -25,9 +25,21 @@
             <argument name="router" xsi:type="object">Magento\WebapiAsync\Controller\VirtualType\Rest\Router</argument>
         </arguments>
     </virtualType>
-    <type name="Magento\WebapiAsync\Controller\Rest\Asynchronous\InputParamsResolver">
+    <virtualType name="Magento\WebapiAsync\Controller\VirtualType\InputParamsResolver" type="Magento\WebapiAsync\Controller\Rest\Asynchronous\InputParamsResolver">
         <arguments>
             <argument name="requestValidator" xsi:type="object">Magento\WebapiAsync\Controller\VirtualType\Rest\RequestValidator</argument>
+            <argument name="isBulk" xsi:type="boolean">true</argument>
         </arguments>
-    </type>
+    </virtualType>
+    <virtualType name="Magento\WebapiAsync\Controller\Rest\VirtualType\AsynchronousBulkRequestProcessor" type="Magento\WebapiAsync\Controller\Rest\AsynchronousRequestProcessor">
+        <arguments>
+            <argument name="inputParamsResolver" xsi:type="object">Magento\WebapiAsync\Controller\VirtualType\InputParamsResolver</argument>
+            <argument name="processorPath" xsi:type="const">Magento\WebapiAsync\Controller\Rest\AsynchronousRequestProcessor::BULK_PROCESSOR_PATH</argument>
+        </arguments>
+    </virtualType>
+    <virtualType name="Magento\WebapiAsync\Controller\Rest\VirtualType\AsynchronousBulkSchemaRequestProcessor" type="Magento\WebapiAsync\Controller\Rest\AsynchronousSchemaRequestProcessor">
+        <arguments>
+            <argument name="processorPath" xsi:type="const">Magento\WebapiAsync\Controller\Rest\AsynchronousSchemaRequestProcessor::BULK_PROCESSOR_PATH</argument>
+        </arguments>
+    </virtualType>
 </config>
